last semestr commit

+++ GoldenTriangle/MathMehods.h
@@ -11,6 +11,6 @@
 @interface MathMehods : NSObject
 
 //kernel must be implement in same class and have same mode like f
-+(CGFloat*)solveEquationVolteraWithKernel:(SEL)kernel f:(SEL)f selectorTarget:(id)selTarget isStatic:(BOOL)isStatic withEndPoint:(float)point lambda:(float)lambda andStep:(float)h;
++(double*)solveEquationVolteraWithKernel:(SEL)kernel f:(SEL)f selectorTarget:(id)selTarget isStatic:(BOOL)isStatic withEndPoint:(double)point lambda:(double)lambda andStep:(double)h;
 
 @end
